% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mksparse.R
\name{addMissing}
\alias{addMissing}
\title{Function to add missingness}
\usage{
addMissing(cmapDS, cmapFold, percentMissing)
}
\arguments{
\item{percentMissing}{percent of missingness to add}
}
\value{
matrix of data with missingness added
}
\description{
Function to add missingness
}
