% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mksparse.R
\name{ckCmapFolds}
\alias{ckCmapFolds}
\title{Function to create folds of cmap dataset}
\usage{
ckCmapFolds(cmapFold, nfolds, oneCellPercent, onePertPercent)
}
\arguments{
\item{cmapFold}{cmap fold dataset from mkCmapFold}
}
\value{
bool does the fold dataset conform to restrictions?
}
\description{
Function to create folds of cmap dataset
}
