% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{cmapBoxPlot}
\alias{cmapBoxPlot}
\title{Function to plot a boxplot of pertubagen class}
\usage{
cmapBoxPlot(expFile, pertFile, geneName)
}
\value{
plotly plot
}
\description{
Function to plot a boxplot of pertubagen class
}
