% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{cmapDoseExtract}
\alias{cmapDoseExtract}
\title{Function to generate data list from metadata objects}
\usage{
cmapDoseExtract(cell, pert, sigInfo, gctxFile, phase = 3)
}
\arguments{
\item{cell}{cmap cell}

\item{pert}{cmap perturbation}

\item{sigInfo}{data for the gctx file to subset}
}
\value{
dataframe of genesxperturbations
}
\description{
Function to generate data list from metadata objects
}
