% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{cmapDT}
\alias{cmapDT}
\title{Function to return a datatable of pertubagen class}
\usage{
cmapDT(
  expFile,
  pertFile,
  inhibit = FALSE,
  keep = 100,
  filterPert = TRUE,
  subClass = NA,
  geneName
)
}
\arguments{
\item{inhibit}{is the direction to sort the table}

\item{keep}{is the nmber of results to display}
}
\value{
datatable
}
\description{
Function to return a datatable of pertubagen class
}
