% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{cmapExtreme}
\alias{cmapExtreme}
\title{Function to isolate top and bottom perturbagens for a particular gene and cell}
\usage{
cmapExtreme(cellName, geneName, sigInfoFile, gctxFile, phase = 3)
}
\arguments{
\item{cellName}{name of cmap cell}

\item{geneName}{gene of interest}

\item{gctxFile}{filepath for a gctx file to subset}

\item{phase}{which phase of data are we using}
}
\value{
a list of dataframes cellsxgenesxperturbations
}
\description{
Function to isolate top and bottom perturbagens for a particular gene and cell
}
