% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{cmapHistPlot}
\alias{cmapHistPlot}
\title{Function to plot a histogram of pertubagen class}
\usage{
cmapHistPlot(expFile, pertFile, subClass = NA, geneName)
}
\arguments{
\item{expFile}{name of cmap extreme data export}

\item{pertFile}{pertubagen export csv}

\item{subClass}{subset of pertubagen file to keep}
}
\value{
plotly plot
}
\description{
Function to plot a histogram of pertubagen class
}
