% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{cmapTox}
\alias{cmapTox}
\title{Function to get a measure of cell viability based on the achilles model}
\usage{
cmapTox(myvec)
}
\arguments{
\item{myvec}{a vector of gene expression from l1000}
}
\value{
a measure of cell viability based on the achilles model
}
\description{
Function to get a measure of cell viability based on the achilles model
}
