% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{getBRDS}
\alias{getBRDS}
\title{Function to get broad ids from perturbagen name}
\usage{
getBRDS(pertName, sigInfo, phase = 3)
}
\arguments{
\item{pertName}{the name of the perturbagen}

\item{sigInfo}{sig file data}

\item{phase}{which phase of data are we using}
}
\value{
a vector of broad ids
}
\description{
Function to get broad ids from perturbagen name
}
