% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{getDosage}
\alias{getDosage}
\title{Function to get broad ids from perturbagen name}
\usage{
getDosage(sigID, sigInfo, phase = 3)
}
\arguments{
\item{sigID}{the name of the perturbagen}

\item{sigInfo}{sig file data}
}
\value{
a vector of broad ids
}
\description{
Function to get broad ids from perturbagen name
}
