% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{getDosagePipeline}
\alias{getDosagePipeline}
\title{Function to run dosage pipeline}
\usage{
getDosagePipeline(pertName, sigInfo, gctxFileLocation, cell, curphase = 3)
}
\value{
a list of data
}
\description{
Function to run dosage pipeline
}
