% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{getPhase1and2}
\alias{getPhase1and2}
\title{Function to run dosage pipeline on phase 1 and 2}
\usage{
getPhase1and2(
  pertName,
  sigInfo1,
  sigInfo2,
  gctxFile1Location,
  gctxFile2Location,
  cell
)
}
\arguments{
\item{sigInfo1}{sig info for phase 1}

\item{sigInfo2}{sig info for phase 1}

\item{gctxFile1Location}{gctx for phase 1}

\item{gctxFile2Location}{gctxfor phase 2}
}
\value{
a list of data
}
\description{
Function to run dosage pipeline on phase 1 and 2
}
