% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{harmonizeGene}
\alias{harmonizeGene}
\title{Function to get broad ids from perturbagen name}
\usage{
harmonizeGene(geneP3)
}
\arguments{
\item{geneP3}{the name of the gene in phase3}
}
\value{
a phase1/2 gene name
}
\description{
Function to get broad ids from perturbagen name
}
