% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{harmonizeGenes}
\alias{harmonizeGenes}
\title{Function to get harmonize genenames}
\usage{
harmonizeGenes(genes)
}
\arguments{
\item{genes}{vector of phase3 gene names}
}
\value{
a vector of phase1/2 gene names
}
\description{
Function to get harmonize genenames
}
