% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCmap.R
\name{labelSubset}
\alias{labelSubset}
\title{Function to classify points}
\usage{
labelSubset(dataOb, quorumPercent = 50)
}
\arguments{
\item{quorumPercent}{the percent of data above tox threshhold to keep any}
}
\value{
dataframe discribing points
}
\description{
Function to classify points
}
