% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{mergeBroad}
\alias{mergeBroad}
\title{Function to combine data across broad ids}
\usage{
mergeBroad(broadIDS, cell, sigInfo, gctxFile, phase = 3)
}
\arguments{
\item{broadIDS}{character vector of Broad ids}
}
\value{
a dataframe of gene expression for 1 cell mult Broad ids
}
\description{
Function to combine data across broad ids
}
