% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{mergeDosageExtracts}
\alias{mergeDosageExtracts}
\title{Function to merge 2 data extracts}
\usage{
mergeDosageExtracts(p1, p2)
}
\arguments{
\item{p1}{output from dosage pipeline}

\item{p2}{output from dosage pipeline}
}
\value{
a list of data

list of data
}
\description{
Function to merge 2 data extracts
}
