% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mksparse.R
\name{mkCmapFold}
\alias{mkCmapFold}
\title{Function to create folds of cmap dataset}
\usage{
mkCmapFold(cmapDS, nfolds, oneCellPercent, onePertPercent)
}
\arguments{
\item{cmapDS}{cmap dataset array cellsxgenesxperts}

\item{nfolds}{number of folds to create}

\item{oneCellPercent}{numeric percent of cells that a fold can have}

\item{onePertPercent}{numeric percent of perturbagens a fold can have}
}
\value{
foldDS array conforming to cmapDS that has fold information
}
\description{
Function to create folds of cmap dataset
}
