% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larimaR.R
\name{plotCmapDosage}
\alias{plotCmapDosage}
\title{Function to Plot Cell viability by dosage}
\usage{
plotCmapDosage(dataOb)
}
\value{
plot of cell viability by dosage
}
\description{
Function to Plot Cell viability by dosage
}
