% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCmap.R
\name{plotCmapFilter}
\alias{plotCmapFilter}
\title{Function to Plot Cell viability by dosage Filtered}
\usage{
plotCmapFilter(dataOb, quorumPercent = 0.5)
}
\value{
plot of cell viability by dosage
}
\description{
Function to Plot Cell viability by dosage Filtered
}
