% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{stouffer}
\alias{stouffer}
\title{Function to weight signatures if multiple signatures found}
\usage{
stouffer(z_scores, weights)
}
\arguments{
\item{z_scores}{a data frame of expression data}

\item{weights}{a vector of weights}
}
\value{
vector of weighted gene expression data
}
\description{
Function to weight signatures if multiple signatures found
}
