% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{weight_signature}
\alias{weight_signature}
\title{Function to generate weights for signatures if multiple signatures found}
\usage{
weight_signature(exprs_df)
}
\arguments{
\item{exprs_df}{a data frame of expression data}
}
\value{
vector of weights
}
\description{
Function to generate weights for signatures if multiple signatures found
}
