% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdataset.R
\name{cmapRdataExtract}
\alias{cmapRdataExtract}
\title{Function to generate data list from metadata objects}
\usage{
cmapRdataExtract(cellDF, pertDF, sigInfoFile, gctxFile, phase = 3)
}
\arguments{
\item{cellDF}{dataframe subset of cmap cell data}

\item{pertDF}{dataframe subset of cmap perturbation data}

\item{gctxFile}{filepath for a gctx file to subset}

\item{phase}{which phase of data are we using}
}
\value{
a list of dataframes cellsxgenesxperturbations
}
\description{
Function to generate data list from metadata objects
}
